NQL_AI_Agent/pyproject.toml
[tool.poetry]
name = "NQL_AI_Agent"
version = "0.1.0"
description = ""
authors = ["qwukker <minor7addfloortom@gmail.com>"]
readme = "README.md"
packages = [{include = "NQL_AI_Agent"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.1"
psycopg2-binary = "^2.9.8"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
matplotlib = "^3.8.2"
pandas = "^2.2.0"
tabulate = "^0.9.0"
websockets = "^13.0.1"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "NQL_AI_Agent.main:main"
start-server = "NQL_AI_Agent.server:main"  